fix: per-locale gate projects + text-extent detection (Chip review)
Two review findings, both real:
- es-419 alone is not a safe maximum — hundreds of pt-BR strings run
  longer. The gate now runs one project per supported non-English locale
  (es-419, pt-BR, es-AR).
- a full-width nowrap child inside a clipping parent ends its box exactly
  at the clip edge while its text keeps going, so the box-edge comparison
  missed it. The detector (now shared in overflow-check.ts) measures text
  extent via scroll size instead, and synthetic self-tests in
  overflow-detector.spec.ts pin every catch/ignore shape.

The stronger detector immediately caught two live landing overflows —
the FAQ heading (PERGUNTAS FREQUENTES, 401px into 320) and the security
section heading (SEGURANÇA., 403px into 320) — both text-heading on
mobile; dropped to the text-heading-xl token, md+ unchanged.
